Overhaul font API so it's now happily object-oriented.
This has also meant introducing some changes to the TextureResource class - flags are now required for specifying whether it's using nearest filtering or not, if mipmaps should be generated and whether or not the background should be discarded. I feel like using flags here is a little friendlier than what we had before.

Discard flag is eeehh. Maybe revisit this later.

I also threw out some debugging crap that was left-over which we didn't really need anymore. In the long term a lot of this will be moved into ImGui.